+++ index.d.ts
@@ -1529,3 +1529,10 @@ export function takeWhile(fn: Predicate<string>, iterable: string): string;
 export function takeWhile(fn: Predicate<string>): (iterable: string) => string;
 export function takeWhile<T>(fn: Predicate<T>, iterable: readonly T[]): T[];
 export function takeWhile<T>(fn: Predicate<T>): (iterable: readonly T[]) => T[];
+
+/**
+ * It returns `true` if property `prop` in `obj1` is equal to property `prop` in `obj2` according to `R.equals`.
+ */
+export function eqProps<T, U>(prop: string, obj1: T, obj2: U): boolean;
+export function eqProps<P extends string>(prop: P): <T, U>(obj1: Record<P, T>, obj2: Record<P, U>) => boolean;
+export function eqProps<T>(prop: string, obj1: T): <U>(obj2: U) => boolean;